+++ analysis/studyTasks/02_cohortDiagnostics.R
@@ -51,8 +51,7 @@ outputFolder <- here::here("results") %>%
   fs::dir_create()
 
 ## Add study variables or load from settings
-diagCohorts <- getCohortManifest() %>%
-  dplyr::filter(type %in% c("target", "outcome"))
+diagCohorts <- getCohortManifest()
 
 
 # E. Script --------------------