+++ extras/executeStudy.R
@@ -13,7 +13,7 @@ source(here::here("analysis/private/_executeStudy.R"))
 # C. Variables -----------------------
 
 ## Add database name
-configBlock <- "synpuf"
+configBlock <- "[block]"
 
 ## Path to tasks folder
 studyTaskFolder <- here::here("analysis/studyTasks")
@@ -28,17 +28,6 @@ runStudyTask(file = studyTaskFiles[1], configBlock = configBlock)
 ## Task 2: Cohort Diagnostics
 runStudyTask(file = studyTaskFiles[2], configBlock = configBlock)
 
-## Task 3: Baseline Characteristics
-runStudyTask(file = studyTaskFiles[3], configBlock = configBlock)
-
-# ## Task 4: 
-# runStudyTask(file = studyTaskFiles[4], configBlock = configBlock)
-# 
-# ## Task 5: 
-# runStudyTask(file = studyTaskFiles[5], configBlock = configBlock)
-# 
-## Task 6: Zip results
-runStudyTask(file = studyTaskFiles[4], configBlock = configBlock)