+++ NEWS.md
@@ -1,4 +1,10 @@
+# bayerThrombosis v0.0.2
+
+* Added `renv` 
+* Removed unnecessary files
+* Updated `.gitignore` file
+
 # bayerThrombosis v0.0.1
 
 * Added JSON cohort definitions
-* Added code to build cohorts and run CohortDiagnostics
+* Added code to build cohorts and run `CohortDiagnostics`